Make lose screen and win screen pretty
The change score functionality and lose screen don't work together so I
need to find a way to fix that maybe with a isNewGame useState constant
but I tried that briefly and it didn't work so I might be better off by
trying something else.

Besides that the most important thing to implement is the function for
the restart button.

Then I just need to make the website responsive for mobile and that
should be it.

After all of that I might come back later to this to add random fetched
movies (unfortunately the tmdb api doesn't have any specific endpoint
for random movies query so it will be a little more laborious than
normal)